Fix method getClosestDistanceToCubsKnown
This method found the closest cub that is visible in the bear's FOV and then added to the provoke system based on how close it was, making the bear more angry if its own cub was close to it. The method is supposed to check for the closest distance between any cub and player in the bear's FOV, which it should now do.